+++ executables/lte-uesoftmodem.c
@@ -73,8 +73,7 @@ pthread_cond_t nfapi_sync_cond;
 pthread_mutex_t nfapi_sync_mutex;
 int nfapi_sync_var=-1; //!< protected by mutex \ref nfapi_sync_mutex
 
-
-uint16_t sf_ahead=4;
+int sf_ahead = 4;
 int tddflag;
 char *emul_iface;
