+++ openair1/SIMULATION/NR_PHY/ulsim.c
@@ -106,9 +106,9 @@ RAN_CONTEXT_t RC;
 char *uecap_file;
 int32_t uplink_frequency_offset[MAX_NUM_CCs][4];
 
-uint16_t sf_ahead=4 ;
+int sf_ahead = 4;
 int slot_ahead=6 ;
-uint16_t sl_ahead=0;
+int sl_ahead = 0;
 double cpuf;
 //uint8_t nfapi_mode = 0;
 uint64_t downlink_frequency[MAX_NUM_CCs][4];
